Kill QMHPluginInterface and make the plugin interface QObject based.
QMHPluginInterface is not a QObject. Our plugins are QObjects though.
Unfortunately, this means that we have to define common properties as
Q_PROPERTY in each and every plugin. This is quite error prone and
unnatural.

Solution is to make QMHPlugin (which is a QObject) our main plugin
interface. With this change, common properties can be defined in the
QMHPlugin and all plugins will automatically get these properties.

A tricky problem is that the moc of QMHPlugin will need to be used by the
plugins.
1. Pass -Wl,-export-dynamic or -rdynamic to gcc so that the symbols in
ELF binary are exported. i.e QMHPlugin's moc code symbols will now be exported
and can be used the plugins. However, this solution is linux/ELF only.
2. Make each plugin have it's own copy of the vtable of QMHPlugin. This
works but a bit scary.
3. Compile QMHPlugin's moc into a separate shared object. This is the approach
that is taken by this commit.