Rearchitect core and skin interaction.
From here on, think of the c++ side as a place for engines. QML
shouldn't have to rely on the c++ side for anything other than
"information/data". For example, things like visualElement/Properties,
have no place in the c++ side. All this can be done in the QML. I am
making the following changes:
0. Backend won't load any engines on startup.
1. QStringList Backend::loadEngines(QStringList engineWhitelist,
engineBlacklist) - will load specific engines. Called from QML. It
load plugins in white list. If white list is empty, it loads all
plugins ignoring the items in backlist. Returns list of loaded
plugins. The skin now control what plugins are loaded. It can say
loadEngine("music") to load just the music plugin and nothing else
even if other engines are present. loadEngine("", "radio") will load
everything other than radio.
2. QML side will use the result of above step to determine what
windows to create. It can thus figure if music has been compiled out,
for example.
3. visualElement/Properties is a concept in the QML side. I will
remove component/Engine altogether because we should be able to show
in the menu without having to load the actual QML (which is not the
case now). Instead we will have addToRootMenu(name, visualElement,
cppengine). Later we will need submenu items and such to be added
here. We can overengineer as we go :P
4. Remove runtime.xxxEngine.