Backend is no more a singleton.
There is no reason it should be. MediaScanner is a singleton instead.
MediaScanner needs to be singleton only for the QML type MediaModel
to be able to connect to scanner updates.

Ideally, if there was a way for QML types to get to the context, we
could get away with it being a non-singleton too.

Note that it makes sense to have multiple Backends anyway since we
might want to stream through multiple cores :P