Create model role mapping on startup
We used to create name->role mapping on the fly. This was not a problem
because mediamodel was tied to a specific database table. However, for
things like a playlist that can possibly contain items of different
types, this dynamic name->role mapping does not work since it assumed
that only one 'table' was active at a time.

Fix is to create the name->role mapping as the media parsers get registered.