Implement auto-forward, auto-backward in the media model
When navigating videos, the season field is "null" for movies. This
means that when having a structure such as "serial|season|title", the
user will be starting at a view that has a single empty item. To avoid
this, we auto enter the season item.

Yes, this is an ugly quick hack for 0.6 release.

071fea9ea7f9e0c3fa162b1375af53af82d2635c was pushed by mistake (instead
of this commit)