heavy prefix refactoring! very sorry about this huge commit, but I had a very bad happening with 'rm -rf' just before pushing my branch
Use qmake PREFIX=/path to set the install prefix
As a developer (in-source) you have to use PREFIX=$PWD/hub !! to get the
old behavior, dont hate me for this :-)

For runtime config use '-skins-path' or 'export QMH_SKINS_PATH' to
adjust the lookup. This is enabled for 'skins', 'keymaps' and 'apps'.