Fix my breakage after upgrading to latest json code.
The latest JSON code parses numbers as doubles as qlonglong.
This means that our slot invocation code now has to handle the
conversion to compatible types (for example, slots that take
an int argument).

The correct fix is to make RpcConnection::handleRpcCall more robust.
But a fix that we can live with for now is to just make all our slots
as qlonglong.

Also, add a comment to highlight limitation of our rpc call code.