Rework media player.
This commit enhances various aspects of our MediaBackend code design
1. VideoPlayer is now an installed module. Currently, we have
implementations for mobility and OverlayMode. We can add more player
types and switch the module import path in C++, so that the QMHPlayer.qml
picks up the correct player.

As can be seen, we can now use aliases to acheive two way bindings. In
general, this should be the suggested approach to build QML components
that have the same API but have a different implementation based on
platform/runtime configuration.

2. VideoPlayerInterface just defines the interface. It is not used by the
anywhere and exists merely for documentation purposes.

3. mediaPlayer is not created by C++. Instead it is registered as a type
(OverlayModeMediaPlayer) which the QML can then use for it's message
passing. This is a pattern that we should be suggesting - Don't create
objects in C++ side, if one can avoid it.